requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch
torchvision
torchaudio
transformers==4.44.0
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-docx==1.1.2
streamlit==1.38.0
opencv-python-headless>=4.10.0
pillow==10.4.0
pandas==2.2.2
fuzzywuzzy[python-levenshtein]==0.18.0
python-multipart==0.0.9
requests==2.32.3
watchdog==4.0.1
pymupdf
